Add __main__ to match with `flet` command
Many Python developers expect that the commands that can be run with `python -m packagename ...` will match the commands that can be run with `packagename ...`. Matching them is also convenient because it allows you to run commands externally with `path/to/env/of/python -m packagename ...` without the environment being applied.
In the case of flet, there's no particular reason not to include `__main__.py`, so I include it.